requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.2
openai==1.35.14
google-api-python-client==2.134.0
google-auth==2.30.0
python-multipart==0.0.9
